requirements.txt
networkx>=3.0
numpy>=1.20
scikit-learn>=1.0
sentence-transformers>=2.2
matplotlib>=3.5
scipy>=1.9
openai>=1.0
pyyaml>=6.0
# Optional — install for Bayesian optimization backend:
# scikit-optimize>=0.9
# Optional — install for TextWorld integration:
# textworld>=1.5
# Optional — install for CMA-ES if not using built-in implementation:
# cma>=3.3
